+++ src/index.js
@@ -1,11 +1,13 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
+import GlobalStyle from "./components/shared/GlobalStyle";
 import App from "./app/app";
 
 ReactDOM.render(
   <React.StrictMode>
     <Router>
+      <GlobalStyle />
       <App />
     </Router>
   </React.StrictMode>,
