+++ src/styles/fonts/font.css
@@ -1,0 +1,42 @@
+@font-face {
+  font-family: "Spoqa Han Sans";
+  font-weight: 200;
+  src: url("../fonts/SpoqaHanSansNeo-Thin.woff") format("woff"),
+    url("../fonts/SpoqaHanSansNeo-Thin.ttf") format("truetype"),
+    url("../fonts/SpoqaHanSansNeo-Thin.otf") format("opentype");
+}
+
+@font-face {
+  font-family: "Spoqa Han Sans";
+  font-weight: 300;
+  src: url("../fonts/SpoqaHanSansNeo-Light.woff") format("woff"),
+    url("../fonts/SpoqaHanSansNeo-Light.ttf") format("truetype"),
+    url("../fonts/SpoqaHanSansNeo-Light.otf") format("opentype");
+}
+
+@font-face {
+  font-family: "Spoqa Han Sans";
+  font-style: normal;
+  font-weight: 400;
+  src: url("../fonts/SpoqaHanSansNeo-Regular.woff") format("woff"),
+    url("../fonts/SpoqaHanSansNeo-Regular.ttf") format("truetype"),
+    url("../fonts/SpoqaHanSansNeo-Regular.ttf") format("opentype");
+}
+
+@font-face {
+  font-family: "Spoqa Han Sans";
+  font-style: medium;
+  font-weight: 500;
+  src: url("../fonts/SpoqaHanSansNeo-Medium.woff") format("woff"),
+    url("../fonts/SpoqaHanSansNeo-Medium.ttf") format("truetype"),
+    url("../fonts/SpoqaHanSansNeo-Medium.otf") format("opentype");
+}
+
+@font-face {
+  font-family: "Spoqa Han Sans";
+  font-style: bold;
+  font-weight: 600;
+  src: url("../fonts/SpoqaHanSansNeo-Bold.woff") format("woff"),
+    url("../fonts/SpoqaHanSansNeo-Bold.ttf") format("truetype"),
+    url("../fonts/SpoqaHanSansNeo-Bold.otf") format("opentype");
+}
